Breakdown some Client files
