+++ Client/Client/StepperView.swift
@@ -1,0 +1,9 @@
+import SwiftUI
+
+struct StepperView: View {
+    @State var value: Int = 0
+
+    var body: some View {
+        Stepper("Stepper", value: $value)
+    }
+}
